+++ pyam/logging.py
@@ -7,9 +7,9 @@ logger = logging.getLogger(__name__)
 
 @contextmanager
 def adjust_log_level(logger="pyam", level="ERROR"):
+    """Context manager to change log level"""
     if isinstance(logger, str):
         logger = logging.getLogger(logger)
-    """Context manager to change log level"""
     old_level = logger.getEffectiveLevel()
     logger.setLevel(level)
     yield