+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+# Next Release
+- [#546](https://github.com/IAMconsortium/pyam/pull/546) Fixed logging for recursive aggregation
+
 # Release v0.12.0
 
 ## Highlights